requirements.txt
python-telegram-bot==20.7
APScheduler==3.10.4
tzlocal==5.2
pybit==5.7.0
pycryptodome==3.20.0
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
mysql-connector-python==8.2.0
perplexityai==0.20.0
numpy==1.26.4

